Muevo cambios en forms de grupos a otro archivo
